Lift the modal dialog keyboard behavior into a shared hook
Escape, Enter and the Tab focus trap were written inline in PositronModalDialog.
PositronDynamicModalDialog is about to need the same three, so they move into
useModalDialogKeyboard, which takes its keydowns from whatever renderer the
dialog is mounted in rather than knowing which one that is.

The logic is carried over as it was, with two exceptions that the next commits
and the comments here call out: the Tab trap changes how it decides what is
focusable, and Escape and Enter now step aside for a quick pick reparented into
the dialog.
